+++ custom_components/metlink_explorer/sensor.py
@@ -264,7 +264,7 @@ class MetlinkRouteSensor(CoordinatorEntity, SensorEntity):
             "name": f"{transportation_name} Route {route_short_name}",
             "manufacturer": "Metlink",
             "model": transportation_name,
-            "sw_version": "0.4.14",
+            "sw_version": "0.4.15",
         }
 
     @property
@@ -358,7 +358,7 @@ class MetlinkDirectionSensor(CoordinatorEntity, SensorEntity):
             "name": f"{transportation_name} Route {route_short_name}",
             "manufacturer": "Metlink",
             "model": transportation_name,
-            "sw_version": "0.4.14",
+            "sw_version": "0.4.15",
         }
 
     @property